+++ python/creation.cpp
@@ -16,24 +16,24 @@ void init_creation(py::module &m) {
       "empty",
       [](const std::vector<Dim> &dims, const std::vector<scipp::index> &shape,
          const units::Unit &unit, const py::object &dtype,
-         const bool with_variance) {
+         const bool with_variances) {
         const auto dtype_ = scipp_dtype(dtype);
         py::gil_scoped_release release;
         return variable::empty(Dimensions(dims, shape), unit, dtype_,
-                               with_variance);
+                               with_variances);
       },
       py::arg("dims"), py::arg("shape"), py::arg("unit") = units::one,
-      py::arg("dtype") = py::none(), py::arg("with_variance") = std::nullopt);
+      py::arg("dtype") = py::none(), py::arg("with_variances") = std::nullopt);
   m.def(
       "ones",
       [](const std::vector<Dim> &dims, const std::vector<scipp::index> &shape,
          const units::Unit &unit, const py::object &dtype,
-         const bool with_variance) {
+         const bool with_variances) {
         const auto dtype_ = scipp_dtype(dtype);
         py::gil_scoped_release release;
         return variable::ones(Dimensions(dims, shape), unit, dtype_,
-                              with_variance);
+                              with_variances);
       },
       py::arg("dims"), py::arg("shape"), py::arg("unit") = units::one,
-      py::arg("dtype") = py::none(), py::arg("with_variance") = std::nullopt);
+      py::arg("dtype") = py::none(), py::arg("with_variances") = std::nullopt);
 }